+++ pom.xml
@@ -26,8 +26,7 @@
 
 	<modules>
 		<module>spring-cloud-task-core</module>
+		<module>spring-cloud-task-samples</module>
 	</modules>
-
-
 
 </project>
