+++ spring-cloud-task-core/src/main/java/org/springframework/cloud/task/configuration/DefaultTaskConfigurer.java
@@ -14,23 +14,31 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.task.config;
+package org.springframework.cloud.task.configuration;
 
+import org.springframework.cloud.task.repository.LoggerTaskRepository;
 import org.springframework.cloud.task.repository.TaskRepository;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
 
 /**
- * Provides a strategy interface for providing configuration
- * customization to the task system.
- *
+ * If no TaskConfigurer is present this configuration will be used.
  * @author Glenn Renfro
  */
-public interface TaskConfigurer {
+@Configuration
+public class DefaultTaskConfigurer {
 
-	/**
-	 * Create a Task Repository for the Task.
-	 *
-	 * @return A TaskRepository
-	 */
-	public TaskRepository taskRepository();
+	@Bean
+	@Scope("prototype")
+	public TaskHandler taskHandler() {
+		return new TaskHandler();
+	}
+
+
+	@Bean
+	public TaskRepository taskRepository() {
+		return new LoggerTaskRepository();
+	}
 
 }
