+++ spring-cloud-task-core/src/main/java/org/springframework/cloud/task/configuration/TaskHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.task.config;
+package org.springframework.cloud.task.configuration;
 
 import java.util.ArrayList;
 import java.util.Date;
